# src/main/resources/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Bank Rest",
    "description": "Bank Rest API",
    "contact": {
      "name": "Dmitriy Folomkin",
      "url": "www.folomkin.dev",
      "email": "dmitriy.folomkin@yandex.ru"
    },
    "license": {
      "name": "License of API",
      "url": "API license URL"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8189",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "Bearer Authentication": []
    }
  ],
...
  "paths": {
    "/api/v1/cards/update-status-card-number/{cardNumber}": {
      "put": {
        "tags": [
          "Карты"
        ],
        "summary": "Обновление статуса карты по номеру",
        "description": "Укажите последние 4 цифры номера карты и одно из доступных значений статуса. Статусы: ACTIVE - Активна, BLOCKED - Заблокирована, ISEXPIRATEDDATE - Истек срок действия",
        "operationId": "updateCardStatusByNumber",
        "parameters": [
          {
            "name": "cardNumber",
            "in": "path",
            "description": "Номер карты",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cards/update-status-card-id/{cardId}": {
      "put": {
        "tags": [
          "Карты"
        ],
        "summary": "Обновление статуса карты по id",
        "description": "Укажите id карты и одно из доступных значений статуса. Статусы: ACTIVE - Активна, BLOCKED - Заблокирована, ISEXPIRATEDDATE - Истек срок действия",
        "operationId": "updateCardStatusById",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "description": "Идентификатор карты",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cards/create-card": {
      "post": {
        "tags": [
          "Карты"
        ],
        "summary": "Создание карты",
        "description": "Форма создания карты. Статус карты по умолчанию \"Активна\". Необходимо указать id пользователя для которого создается карта или 0 для текущего пользователя",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cards/filter-cards": {
      "get": {
        "tags": [
          "Карты"
        ],
        "summary": "Получение списка всех карт постранично",
        "operationId": "getCardsPages",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "description": "Номер страницы с результатом",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Количество выводимых карт на странице. Минимум 1, максимум 50",
            "required": false,
            "schema": {
              "maximum": 50,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Поле сортировки",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "owner",
            "in": "query",
            "description": "Email пользователя",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedModelCardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cards/cards-userid/{userId}": {
      "get": {
        "tags": [
          "Карты"
        ],
        "summary": "Получение списка всех карт пользователя",
        "operationId": "getCardsByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "Id пользователя",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cards/cards-blockrequest": {
      "get": {
        "tags": [
          "Карты"
        ],
        "summary": "Получение списка карт с запросом на блокировку",
        "operationId": "getCardsByBlockRequest",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cards/card-number/{number}": {
      "get": {
        "tags": [
          "Карты"
        ],
        "summary": "Получение карты по номеру",
        "description": "Укажите последние 4 цифры номера карты",
        "operationId": "getCardByNumber",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "description": "Номер карты",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cards/card-id/{id}": {
      "get": {
        "tags": [
          "Карты"
        ],
        "summary": "Получение карты по id",
        "description": "Поиск по id карты",
        "operationId": "getCardById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id карты",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cards/all": {
      "get": {
        "tags": [
          "Карты"
        ],
        "summary": "Получение списка всех карт",
        "description": "Список без пагинации и фильтров",
        "operationId": "getAllCards",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cards/delete-card-number/{number}": {
      "delete": {
        "tags": [
          "Карты"
        ],
        "summary": "Удаление карты по номеру",
        "description": "Укажите последние 4 цифры карты",
        "operationId": "deleteCardByNumber",
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "description": "Номер карты",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cards/delete-card-id/{cardId}": {
      "delete": {
        "tags": [
          "Карты"
        ],
        "summary": "Удаление карты по id",
        "description": "Укажите id карты",
        "operationId": "deleteCardById",
        "parameters": [
          {
            "name": "cardId",
            "in": "path",
            "description": "ID карты",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {
              "type": "object"
            }
          }
        }
      },
      "CardStatusRequest": {
        "required": [
          "cardStatus"
        ],
        "type": "object",
        "properties": {
          "cardStatus": {
            "type": "string",
            "description": "Статус карты",
            "example": "ACTIVE, BLOCKED, ISEXPIRATEDDATE",
            "enum": [
              "ACTIVE",
              "BLOCKED",
              "ISEXPIRATEDDATE",
              "BLOCKREQUEST"
            ]
          }
        },
        "description": "Статус карты"
      },
      "CardResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "encryptedNumber": {
            "type": "string",
            "description": "Номер карты с маской"
          },
          "expirationDate": {
            "type": "string",
            "description": "Дата окончания срока действия карты",
            "format": "date"
          },
          "cardStatus": {
            "type": "string",
            "description": "Статус карты",
            "enum": [
              "ACTIVE",
              "BLOCKED",
              "ISEXPIRATEDDATE",
              "BLOCKREQUEST"
            ]
          },
          "balance": {
            "type": "number",
            "description": "Баланс карты"
          },
          "owner": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "description": "Возвращаемый объект при регистрации карты"
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "description": "Пользователь карты"
      },
...
      "CardRequest": {
        "required": [
          "balance",
          "expirationDate",
          "openNumber",
          "userId"
        ],
        "type": "object",
        "properties": {
          "openNumber": {
            "maxLength": 19,
            "minLength": 19,
            "pattern": "^(\\d{4}(\\s)\\d{4}(\\s)\\d{4}(\\s)\\d{4})$",
            "type": "string",
            "description": "Номер карты"
          },
          "expirationDate": {
            "type": "string",
            "description": "Дата окончание срока действия карты",
            "format": "date"
          },
          "balance": {
            "type": "number",
            "description": "Баланс карты"
          },
          "userId": {
            "type": "integer",
            "description": "ID владельца карты",
            "format": "int64"
          }
        },
        "description": "Форма регистрации карты"
      },
...
      "PageMetadata": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "PagedModelCardResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CardResponse"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}